+++ examples/hobotnica/app.py
@@ -2,7 +2,6 @@ from pathlib import Path
 from typing import List
 
 from aiohttp import web
-from aiohttp_middlewares import cors_middleware, error_middleware
 
 from rororo import BaseSettings, setup_openapi, setup_settings
 from . import views
@@ -16,16 +15,12 @@ def create_app(
 
     return setup_openapi(
         setup_settings(
-            web.Application(
-                middlewares=(
-                    cors_middleware(allow_all=True),
-                    error_middleware(),
-                )
-            ),
+            web.Application(),
             settings,
             loggers=("aiohttp", "aiohttp_middlewares", "hobotnica", "rororo"),
             remove_root_handlers=True,
         ),
         Path(__file__).parent / "openapi.yaml",
         views.operations,
+        cors_middleware_kwargs={"allow_all": True},
     )