+++ examples/hobotnica/views.py
@@ -1,15 +1,22 @@
-import logging
-
 from aiohttp import web
 
 from rororo import openapi_context, OperationTableDef
+from rororo.openapi.exceptions import ObjectDoesNotExist
 from .data import GITHUB_REPOSITORIES
 from .decorators import login_required
-from .exceptions import ObjectDoesNotExist
 
 
-logger = logging.getLogger(__name__)
 operations = OperationTableDef()
+
+
+@operations.register
+@login_required
+async def list_owner_repositories(request: web.Request) -> web.Response:
+    with openapi_context(request) as context:
+        username = context.security["basic"].login
+        return web.json_response(
+            list((GITHUB_REPOSITORIES.get(username) or {}).values())
+        )
 
 
 @operations.register